requirements.txt
langchain>=1.2.0
langchain-google-genai>=4.2.1
langchain-community>=0.4.0
chromadb>=0.6.0
pypdf>=5.0.0
python-dotenv>=1.0.1
streamlit>=1.42.0